MOVEONLY: move package unit tests to their own file
